+++ src/main/java/com/codeborne/selenide/collections/ExactTexts.java
@@ -68,7 +68,7 @@ public class ExactTexts extends CollectionCondition {
   }
 
   @Override
-  public boolean applyNull() {
+  public boolean missingElementSatisfiesCondition() {
     return false;
   }