+++ src/main/java/com/codeborne/selenide/collections/SizeGreaterThan.java
@@ -31,7 +31,7 @@ public class SizeGreaterThan extends CollectionCondition {
   }
 
   @Override
-  public boolean applyNull() {
+  public boolean missingElementSatisfiesCondition() {
     return apply(0);
   }