+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * #1467 Avoid spam in logs when webdriver is already closed  --  see PR #1540
 * #1534 `Or` and `And` conditions work correctly with non-existent element  --  thanks to Evgenii Plugatar for PR #1539
 * `Or` and `And` conditions support PECS principle in ctor, no longer allow empty list in ctor  --  thanks to Evgenii Plugatar for PR #1542
-* #1541 removed deprecated `Condition` method `applyNull()`  --  thanks to Evgenii Plugatar for PR #1544
+* #1541 removed deprecated `Condition` method `applyNull()` and renamed `CollectionCondition` `applyNull()` method  --  thanks to Evgenii Plugatar for PR #1544
 
 ## 5.23.3 (released 19.08.2021)
 * #1528 fix "exe" or "dmg" file download in Chrome  -  see PR #1529